Use board-2.bin file containing firmware data from Windows
The QCA6174 chip, used on the Surface Go and the Surface Laptop 3,
requires firmware files to store important eeprom data.

While the board.bin file used previously, obtained from Killer
Networking, does seem to fix the issues and contain data that is
apparently close enough to get things working, it does not contain data
that is explicitly tailored to the chips used in the Surface Go and
Surface Laptop 3.

The board-2.bin file contains extended information that is not contained
in the generic/default board.bin file. This extended data is tied to the
PCI device vendor/ID combination. Hans de Goede has created a
board-2.bin file that containes this extended data for both the Surface
Go and the Surface Laptop 3, based on firmware files found on Windows.

Thus update the hack replacing the board.bin file to replace the
board-2.bin file with the extended one instead.

Refer to the links below for more information:
 - https://github.com/linux-surface/linux-surface/issues/542
 - https://github.com/linux-surface/linux-surface/issues/41
 - https://github.com/kvalo/ath10k-firmware/pull/11