pkg/arch: Ignore dependencies
The package depends on the linux-surface kernel package. This is
somewhat hacky as the patched kernel is required for the workaround to
work and we use this as mechanism to encourage people to install that.
It is not required during build time.

Installing the patched kernel would require adding the repo and since we
don't need it at packaging time, let's just ignore dependencies and skip
all that.